Database partition mode Part 2 (#6409)
* First pass at segmenting out changes

* Test fixes

* Test fixes

* Test fix

* Work on tests

* Fix test

* Test fix

* Test fix

* Fixes

* Test fixes

* Test fix

* License headers

* Test fix

* Add changelog

* Address review comments

* Review comments

* Fixes

* Work on tests

* Address review comments

* Fix

* Add tests

* Cleanup

* Rename JpaPidValueTuples

* Add changelog

* Spotless

* Resolve compile issues

* Fix signature issue

* Test fix

* Add header

* Spotless

* Test cleanup

* Work on merge

* Work on tests

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/expunge/ExpungeEverythingService.java
@@ -321,10 +321,12 @@ public class ExpungeEverythingService implements IExpungeEverythingService {
 						nativeQuery.setMaxResults(800);
 						List pids = nativeQuery.getResultList();
 
-						nativeQuery = myEntityManager.createQuery("DELETE FROM " + theEntityType.getSimpleName()
-								+ " WHERE " + idProperty + " IN (:pids)");
-						nativeQuery.setParameter("pids", pids);
-						nativeQuery.executeUpdate();
+						if (!pids.isEmpty()) {
+							nativeQuery = myEntityManager.createQuery("DELETE FROM " + theEntityType.getSimpleName()
+									+ " WHERE " + idProperty + " IN (:pids)");
+							nativeQuery.setParameter("pids", pids);
+							nativeQuery.executeUpdate();
+						}
 						return pids.size();
 					});
